Update setting for `max-locals` pylint's rule
`R0914` (`too-many-locals`) will be raised if sum of local variables of
a function and all arguments of the same function is greater then
value of `max-locals`.

Now we allow up to 5 arguments for a function. And we allow up to 10
local variables + arguments for a function. It means if we have 3
arguments, we can define up to 7 local variables.