Don't rewrite exit codes during code assessment
`pylint` is executed after `black`. And if `black` is failed, but
`pylint` is passed, the final result will be passed. It means that the
exit code of the latest tool in the code assessment decides if the
procedure is failed or not.

The described behavior was changed in a way that code assessment will
fail if at least one of the tools is failed. Now, there is a sequence of
failures. And if it is not empty, it contains names of tools which have
found some inconsistencies within a code base.