Implement login test case via API
Login test case was implemented for checking ability login into
user account (with valid credentials) using API.

Opencart website was deployed on remote VM, connection to the site
is not secure. For ability make login with 'https' protocol, was
added parameter 'verify' in thr post method . In this case warning
'InsecureRequestWarning' is shown, for disabling this warning,
'urllib3' library was imported.

In file requirements.txt was added name of 'urllib3' library with
current version. #26