Implement returns test case via API
Returns test case was implemented for checking ability send
returns form using API.

Returns form is sent with post request, which was used in test case.
Opencart website was deployed on remote VM, connection to the site
is not secure. For ability make login with 'https' protocol, was
added parameter 'verify'. In this case warning
'InsecureRequestWarning' is shown, for disabling this warning,
'urllib3' library was imported.