Add review check test case via API
This test case checks ability to write review via API.
In test case using post request to the
http://localhost/index.php?route=product/product/write&product_id=40
we send parameters:
	1."name"
	2."text"
	3."rating"
In the case of a successful passage we expect to see a message on
the title of the page:
"Thank you for your review. It has been submitted to the webmaster
for approval."
In the unsuccessful case, the test case won't pass.