Creating unittests for email generator
Function creating_full_email was covered by unit tests. Unit tests are verifying that function works correctly.
In unit tests were used regular expressions for verifying correct patterns of random email or some parts of the email.
In unit tests were used asserts on the length of emails parts.
In unit tests was used the assert for the absence of special symbols.