Create unitests for creating_full_email function.
The minimum email length is checked. Local part minimum 2 characters + "@ mailinator.com" (15 characters) = 17 characters.
Check for maximum length (254 characters).
Check if there is a domain name.
Check whether there is a "com."
Check for the number of characters "@".
Check if there is a ". @".
Check if the first character is a letter. #47
In code_assesment.sh was added py.test for runing unitests.
In README.md was added information about py.test.